meetIQ-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
google-adk>=0.1.0
openai-whisper
